goods/requirements.txt
Django
psycopg2-binary
djangorestframework
django-rest-swagger
django-cleanup
drf_yasg
Pillow
requests